+++ docs/conf.py
@@ -29,6 +29,9 @@ copybutton_prompt_text = r">>> |\.\.\. |\$ "
 copybutton_prompt_is_regexp = True
 
 html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
+html_logo = "logo_capella.png"
+
 
 html_theme_options = {
     "navigation_depth": 2,